use keys instead of each_key for transforming keys
This reverses rubocop's suggestion to use `#each_key` as the latter
method lazily iterates over an object's keys, not allowing in-place
modification of the object whose keys are being iterated. The `#keys`
method on the other hand creates a new array of the keys which will
work with the current logic. This should probably be refactored
a bit, especailly since newer Ruby versions allow deep key
transformations natively.