bump to latest rubocop
This adds a config setting to keep the dev dependencies specified
in the gemspec file. The `Gemspec/DevelopmentDependencies` cop was
introduced as of rubocop v1.44 and by default enforces that the
dev dependencies should be specified in the Gemfile. There does
not seem to be a consensus in the community as to where these
dependencies sould be stores, however, bundler's author [1]
recommended putting everything in gemspec and this rubygems
issue [2] shows that the maintainers made an argument that
keeping everything in gemspec helps keeps things more DRY.
Setting the style config to `gemspec` keeps the current
style for this gem.

This also locks the version down to the minor level, ensuring that
the latest version of the gem fetched by default via bundle will
not result in a minor version bump, which should keep the cops
stable. This should allow us to upgrade the gem to minor/major
versions as desired by updating the version explicitly.

[1] https://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/
[2] https://github.com/rubygems/rubygems/issues/1104