make util methods private
Updated specs to test the private methods directly, which is typically
considered a smell. In this case we're leaving the specs to test the
implementation of these methods directly as there is still value in
having their logic be covered explicitly. Might come back to rework
the util methods if time permits (e.g extract to a dedicated module
perhaps).